pull blockchain stuff into its own pkg
